@@ -9,23 +9,23 @@
     "postbuild": "node ./build-tools/set_version_in_files.js dist/index.html",
     "prebuild": "npm run lang",
     "prelang": "npm run clean",
-    "analyze": "npx eslint -c .eslintrc.json src/",
+    "analyze": "npx eslint -c eslint.config.js src/",
     "stylelint": "npx stylelint src/scss/**/*.scss"
   },
   "dependencies": {
-    "@parcel/core": "^2.5.0",
+    "@parcel/core": "^2.12.0",
     "@storeon/crosstab": "^1.0.2",
-    "@storeon/localstorage": "^1.4.0",
-    "bootstrap": "^5.1.3",
+    "@storeon/localstorage": "^1.5.0",
+    "bootstrap": "^5.3.3",
     "glightbox": "^3.2.0",
     "i18next": "^23.2.11",
     "idb-keyval": "^6.1.0",
     "image-blob-reduce": "^4.1.0",
     "js-cookie": "^3.0.1",
-    "mobile-drag-drop": "^3.0.0-beta.0",
+    "mobile-drag-drop": "^3.0.0-rc.0",
     "node-sass": "^9.0.0",
-    "preact": "10.17.1",
-    "react-i18next": "^13.0.2",
+    "preact": "^10.24.2",
+    "react-i18next": "^15.0.3",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
     "storeon": "^3.1.5"
@@ -39,24 +39,24 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
+    "@eslint/js": "^9.12.0",
     "@parcel/packager-raw-url": "^2.5.0",
     "@parcel/transformer-image": "^2.5.0",
     "@parcel/transformer-sass": "^2.5.0",
     "@parcel/transformer-webmanifest": "^2.5.0",
     "@size-limit/preset-app": "^8.0.1",
     "babel-plugin-module-resolver": "^5.0.0",
-    "eslint": "^8.13.0",
-    "eslint-config-preact": "^1.3.0",
-    "eslint-plugin-compat": "^4.0.2",
-    "eslint-plugin-jest": "^27.0.3",
+    "eslint": "^9.12.0",
+    "eslint-plugin-compat": "^6.0.1",
     "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "globals": "^15.11.0",
     "parcel": "^2.5.0",
     "prettier": "^3.0.0",
     "sass": "^1.50.1",
-    "size-limit": "^8.0.1",
-    "stylelint": "^15.11.0",
-    "stylelint-config-standard-scss": "^11.0.0"
+    "size-limit": "^11.1.6",
+    "stylelint": "^16.10.0",
+    "stylelint-config-standard-scss": "^13.1.0"
   },
   "alias": {
     "react": "preact/compat",
